requirements.txt
torch
torchvision
numpy
matplotlib
streamlit
pillow
streamlit-drawable-canvas